add redirect to error page in case of edition of non existing or some…

+++ src/main/java/ua/goit/goitnotes/note/NoteController.java
@@ -5,12 +5,7 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.ResponseBody;
+import org.springframework.web.bind.annotation.*;
 import ua.goit.goitnotes.error_handling.ObjectNotFoundException;
 import ua.goit.goitnotes.note.dto.FormattedNote;
 import ua.goit.goitnotes.note.dto.NoteDTO;
@@ -25,7 +20,6 @@ import ua.goit.goitnotes.validation.ValidationService;
 
 import java.util.Comparator;
 import java.util.List;
-import java.util.Objects;
 import java.util.UUID;
 import java.util.stream.Collectors;
 
@@ -64,7 +58,12 @@ public class NoteController {
     @GetMapping(path = "/delete")
     public String delete(@RequestParam(name = "id") UUID uuid) {
         log.info("NoteController.delete().");
-        noteService.delete(uuid);
+        NoteDTO note = noteService.findById(uuid);
+        if (note.getUserName().equals(SecurityContextHolder.getContext().getAuthentication().getName())) {
+            noteService.delete(uuid);
+        } else {
+            return "redirect:/error";
+        }
         return "redirect:/note/list";
     }
